AI-LiteLLM/pyproject.toml
[project]
name = "ai-litellm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "litellm>=1.74.4",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.0",
]
